requirements.txt
openai
pandas
pymongo
matplotlib
opencv-python
python-dotenv
